backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.216.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
playwright==1.40.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
# 画像処理（サイト複製機能用）
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
